Multivariate forward stepwise regression

Multivariate forward stepwise regression evluated by multivariate BIC

mvForwardStepwise(
  exprObj,
  baseFormula,
  data,
  variables,
  criterion = c("BIC", "sum BIC", "AIC", "AICC", "CAIC", "sum AIC"),
  shrink.method = c("EB", "none", "var_equal", "var_unequal"),
  nparamsMethod = c("edf", "countLevels", "lme4"),
  deltaCutoff = 5,
  fastApprox = FALSE,
  verbose = TRUE,
  ...
)

Arguments

exprObj
matrix of expression data (g genes x n samples), or ExpressionSet, or EList returned by voom() from the limma package

baseFormula
specifies baseline variables for the linear (mixed) model. Must only specify covariates, since the rows of exprObj are automatically used as a response. e.g.: ~ a + b + (1|c) Formulas with only fixed effects also work, and lmFit() followed by contrasts.fit() are run.

data
data.frame with columns corresponding to formula

variables
array of variable names to be considered in the regression. If variable should be considered as random effect, use '(1|A)'.

criterion
multivariate criterion ('AIC', 'BIC') or summing score assuming independence of reponses ('sum AIC', 'sum BIC')

shrink.method
Shrink covariance estimates to be positive definite. Using "var_equal" assumes all variance on the diagonal are equal. This method is the fastest because it is linear time. Using "var_unequal" allows each response to have its own variance term, however this method is quadratic time. Using "none" does not apply shrinkge, but is only valid when there are very few responses

nparamsMethod
"edf": effective degrees of freedom. "countLevels" count number of levels in each random effect. "lme4" number of variance compinents, as used by lme4. See description in nparam

deltaCutoff
stop interating of the model improvement is less than deltaCutoff. default is 5

fastApprox
use PCA to transform variables and substantially decrease compute time. This approach is exact when only fixed effects are used, and approximate when random effects are included.

verbose
Default TRUE. Print messages

...
additional arguements passed to logDet

Value

list with formula of final model, and trace of iterations during model selection

Examples

Y = with(iris, rbind(Sepal.Width, Sepal.Length))

# fit forward stepwise regression starting with model: ~1. 
bestModel = mvForwardStepwise( Y, ~ 1, data=iris, variables=colnames(iris)[3:5])

#> Base model: ~ 1

#> 	evaluating: + Petal.Length                   

#> 	evaluating: + Petal.Width                   

#> 	evaluating: + Species                   

#> 
#> Best model delta: -293

#> Add variable to model: Petal.Length

#> Base model: ~ 1 + Petal.Length

#> 	evaluating: + Petal.Width                   

#> 	evaluating: + Species                   

#> 
#> Best model delta: -73

#> Add variable to model: Species

#> Base model: ~ 1 + Petal.Length + Species

#> 	evaluating: + Petal.Width                   

#> 
#> Best model delta: -15

#> Add variable to model: Petal.Width

bestModel

#> 		Multivariate IC forward stepwise regression
#> 
#>   Samples:	 150 
#>   Responses:	 2 
#>   Shrink method: EB 
#>   Criterion:	 BIC 
#>   Iterations:	 3 
#> 
#>   Best model: ~ 1 + Petal.Length + Species + Petal.Width 
#>